# docs/swagger.yaml
basePath: /
definitions:
  handlers.CreateRequestInput:
    properties:
      certificate:
        type: string
      description:
        type: string
      status:
        type: string
      type:
        type: string
    required:
    - description
    - type
    type: object
...
  handlers.UpdateRequestStatusRequest:
    properties:
      request_id:
        type: string
      status:
        type: string
    required:
    - request_id
    - status
    type: object
...
host: 109.73.202.151:8080
info:
  contact: {}
  description: ITaM API
  title: LiveCode API
  version: "1.0"
paths:
...
  /auth/api/create_user_request:
    post:
      consumes:
      - application/json
      description: Создает новый запрос от имени пользователя
      parameters:
      - description: Request data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateRequestInput'
      produces:
      - application/json
      responses:
        "200":
          description: Success message with request ID
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - OAuth2Password: []
      summary: Создать запрос пользователя
      tags:
      - Requests
...
  /auth/api/delete_request:
    delete:
      description: Удаляет запрос по его ID
      parameters:
      - description: Request ID
        in: query
        name: request_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success message
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid request ID
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Request not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - OAuth2Password: []
      summary: Удалить запрос
      tags:
      - Requests
...
  /auth/api/get_all_requests:
    get:
      description: Возвращает список всех запросов текущего пользователя с пагинацией
      parameters:
      - description: User ID
        in: query
        name: user_id
        required: true
        type: string
      - default: 10
        description: Limit
        in: query
        name: limit
        type: integer
      - default: 0
        description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: All requests
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid user ID or pagination parameters
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - OAuth2Password: []
      summary: Получить все запросы пользователя
      tags:
      - Requests
...
  /auth/api/get_request:
    get:
      description: Возвращает список запросов пользователя с пагинацией
      parameters:
      - description: User ID
        in: query
        name: user_id
        required: true
        type: string
      - default: 10
        description: Limit
        in: query
        name: limit
        type: integer
      - default: 0
        description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Request data
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid user ID or pagination parameters
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - OAuth2Password: []
      summary: Получить запросы пользователя
      tags:
      - Requests
...
  /auth/api/update_request_status:
    patch:
      consumes:
      - application/json
      description: Обновляет статус указанного запроса
      parameters:
      - description: Request status update data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateRequestStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Success message
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - OAuth2Password: []
      summary: Обновить статус запроса
      tags:
      - Requests
...
schemes:
- http
securityDefinitions:
  OAuth2Password:
    flow: password
    tokenUrl: /api/login
    type: oauth2
swagger: "2.0"